Remove pending header deps on Interpreters/Set.h
