+++ src/Processors/QueryPlan/CreateSetAndFilterOnTheFlyStep.cpp
@@ -3,6 +3,7 @@
 
 #include <QueryPipeline/QueryPipelineBuilder.h>
 #include <IO/Operators.h>
+#include <Interpreters/SetWithState.h>
 #include <Common/JSONBuilder.h>
 #include <Common/logger_useful.h>
 #include <Core/ColumnWithTypeAndName.h>
